refactor: guards folder structure
