+++ src/private/private.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, UseGuards } from '@nestjs/common';
-import { JwtAuthGuard } from 'src/guards/jwt-auth.guard';
-import { RateLimitGuard } from 'src/guards/rate-limit.guard';
+import { JwtAuthGuard } from 'src/guards/jwt-auth/jwt-auth.guard';
+import { RateLimitGuard } from 'src/guards/rate-limit/rate-limit.guard';
 import { PrivateService } from './private.service';
 
 @UseGuards(JwtAuthGuard)
